+++ src/main/java/com/agentevirtual/repository/ClienteRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {
-    Optional<Cliente> findByIdentificacion(String identificacion);
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {
+    
+	Optional<Cliente> findByIdentificacion(String identificacion);
 }
